Removed structures and added steam and removed plasma.

+++ src/main/java/main/energy/Generator.java
@@ -1,6 +1,6 @@
 package main.energy;
 
-public record Generator(int fluidUsagePerTick,
+public record Generator(double fluidUsagePerTick,
                         int maximumTemperature,
                         double effectivity) {
 }
